requirements.txt
anyio==3.6.2
appdirs==1.4.4
APScheduler==3.6.3
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.1
cachetools==4.2.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
frozendict==2.3.4
h11==0.14.0
html5lib==1.1
httpcore==0.16.3
httpx==0.23.3
idna==3.4
Jinja2==3.1.2
jsonschema==4.17.3
lxml==4.9.2
MarkupSafe==2.1.2
multitasking==0.0.11
numpy==1.24.1
outcome==1.2.0
pandas==1.5.3
pycparser==2.21
pyproject==1.3.1
pyproject-toml==0.0.10
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-telegram-bot==13.7
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rfc3986==1.5.0
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
toml==0.10.2
tornado==6.2
tradingview-ta==3.3.0
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
tzlocal==4.2
Unidecode==1.3.6
urllib3==1.26.14
webencodings==0.5.1
wsproto==1.2.0
yfinance==0.2.9
